+++ packages/searchbar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pikas-ui/searchbar
 
+## 1.1.1
+
+### Patch Changes
+
+- Add width, minWidth and maxWidth props
+- Updated dependencies
+  - @pikas-ui/button@1.0.26
+  - @pikas-ui/textfield@1.0.27
+
 ## 1.1.0
 
 ### Minor Changes
@@ -18,7 +27,6 @@
 
 - Create searchbar component
 - Update typing exported
-- Updated dependencies
 - Updated dependencies
   - @pikas-ui/button@1.0.25
   - @pikas-ui/separator@1.0.25